+++ CHANGELOG.md
@@ -9,3 +9,9 @@
 ## [1.0.2] - 2025-11-09
 - Fixed calibrate_y logic which used the beacon safe home setting and ignored the user's override
 - Added attribution to graphs
+
+## [1.1.0] - 2025-11-16
+- Updated install script to add archival repositories for EOL Debian versions
+- Print head will now be at beacon safe home position after homing
+- Added check to exit compensation mode early if user did not add axis_twist_compensation module to config
+- Updated readme